domain 수정
